+++ src/components/location-finder/SearchAutoSuggest/SearchAutoSuggest.js
@@ -1,6 +1,6 @@
 import React, { Fragment, useState } from 'react';
 // Apollo
-import { useLazyQuery } from '@apollo/react-hooks';
+import { useLazyQuery } from '@apollo/client';
 import { LocationsQuery as LOCATIONS_QUERY } from './SearchAutoSuggest.gql';
 // Redux
 import {