+++ src/utils/filterByOpenNow.js
@@ -1,4 +1,4 @@
-import checkAlertsOpenStatus from './checkAlertsOpenStatus';
+import hasActiveClosing from './hasActiveClosing';
 
 /**
  * Takes an array of locations and reduces the array to only open now Locations.
@@ -13,9 +13,9 @@ function filterByOpenNow(now, locations) {
   const weekday = now.format('ddd');
   const today = now.format();
 
-  return locations.reduce((accumlator, location) => {
+  return locations.reduce((accumulator, location) => {
     // Alerts
-    const alertsOpenStatus = checkAlertsOpenStatus(today, location._embedded.alerts);
+    const isActiveClosing = hasActiveClosing(today, location._embedded.alerts, null);
 
     location.hours.regular.map(hoursItem => {
       // Find today in weekly hours.
@@ -24,19 +24,17 @@ function filterByOpenNow(now, locations) {
         if (
           // Check for not null.
           hoursItem.open !== null && hoursItem.close !== null
-          // Check for alert closings.
-          && alertsOpenStatus
-          // Check for extended closing.
-          && location.open
+          // Check for active closing.
+          && !isActiveClosing
           // Check open/closed hours against now time.
           && hoursItem.open <= nowTime && hoursItem.close >= nowTime
         ) {
           // Add location as open to accumulator.
-          accumlator.push(location);
+          accumulator.push(location);
         }
       }
     });
-    return accumlator;
+    return accumulator;
   }, []);
 }