+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Added tests for LocationAccessibility component.
 * Handle late openings and early closings.
 * Code refactoring and test coverage for closings and regular hours.
+* Upgrades to Apollo Client v3 and removes older apollo client packages.
+* Fixes duplicate render issue causes by older apollo client bugs.
 
 ### v0.1.4 [2020-11-10] Location Finder Improvements
 -----------------------------------------------