+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Code refactoring and test coverage for closings and regular hours.
 * Upgrades to Apollo Client v3 and removes older apollo client packages.
 * Fixes duplicate render issue causes by older apollo client bugs.
+* Locks NextJS to 9.5.5 instead of latest.
 
 ### v0.1.4 [2020-11-10] Location Finder Improvements
 -----------------------------------------------