+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Locks NextJS to 9.5.5 instead of latest.
 * Adds basic test coverage for locations component and mock apollo data.
 * Upgraded DGX React Footer 0.5.7 -> 0.5.8 to fix accessibility issue.
+* Adds "main-content" id to the main HTML element.
 
 ### v0.1.4 [2020-11-10] Location Finder Improvements
 -----------------------------------------------