+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 ### v0.1.5 [?] Location Finder Improvements
 -----------------------------------------------
 * Added tests for LocationAccessibility component.
+* Handle late openings and early closings.
+* Code refactoring and test coverage for closings and regular hours.
+* Upgrades to Apollo Client v3 and removes older apollo client packages.
+* Fixes duplicate render issue causes by older apollo client bugs.
+* Locks NextJS to 9.5.5 instead of latest.
+* Adds basic test coverage for locations component and mock apollo data.
+* Upgraded DGX React Footer 0.5.7 -> 0.5.8 to fix accessibility issue.
+* Adds "main-content" id to the main HTML element.
 
 ### v0.1.4 [2020-11-10] Location Finder Improvements
 -----------------------------------------------